+++ orttraining/orttraining/training_ops/cuda/math/mixed_precision_scale_impl.cu
@@ -5,9 +5,9 @@
 #pragma warning(disable : 4244)
 #endif
 
+#include "mixed_precision_scale_impl.h"
 #include <cuda_fp16.h>
 #include "core/providers/cuda/cu_inc/common.cuh"
-#include "mixed_precision_scale_impl.h"
 
 namespace onnxruntime {
 namespace cuda {
