+++ orttraining/orttraining/training_ops/cuda/math/mixed_precision_scale_impl.h
@@ -3,6 +3,8 @@
 
 #pragma once
 
+#include <cuda_runtime.h>
+
 namespace onnxruntime {
 namespace cuda {
 template <typename SrcT, typename DstT>
