+++ orttraining/orttraining/training_ops/cuda/math/scale.cc
@@ -26,6 +26,15 @@ namespace cuda {
           .InputMemoryType(OrtMemTypeCPUInput, 1),                             \
       Scale<T>);
 
+template <typename ScaleT>
+struct GetScaleValueImpl {
+  void operator()(const Tensor* scale, float& scale_value) const {
+    ORT_ENFORCE(scale->Shape().Size() == 1, "Scale input should have a single value.");
+    scale_value = static_cast<float>(*(scale->template Data<ScaleT>()));
+    ORT_ENFORCE(scale_value != 0.0f, "Scale value must not be 0.");
+  }
+};
+
 template <typename T>
 Scale<T>::Scale(const OpKernelInfo& info) : CudaKernel(info) {
   int64_t scale_down;
