+++ orttraining/orttraining/training_ops/cuda/optimizer/adam.cc
@@ -1,12 +1,13 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+#include "orttraining/training_ops/cuda/optimizer/adam.h"
+#include "orttraining/training_ops/cuda/optimizer/adam_impl.h"
+
 #include "core/providers/cuda/cuda_allocator.h"
 #include "core/providers/cuda/reduction/reduction_functions.h"
 #include "core/providers/cuda/math/binary_elementwise_ops.h"
 #include "orttraining/training_ops/cuda/optimizer/common.h"
-#include "orttraining/training_ops/cuda/optimizer/adam.h"
-#include "orttraining/training_ops/cuda/optimizer/adam_impl.h"
 
 namespace onnxruntime {
 namespace cuda {
