+++ orttraining/orttraining/training_ops/cuda/optimizer/adam_impl.h
@@ -4,6 +4,7 @@
 #pragma once
 
 #include <cstdint>
+#include <cuda_runtime.h>
 
 namespace onnxruntime {
 namespace cuda {
