+++ orttraining/orttraining/training_ops/cuda/optimizer/gradient_control.cc
@@ -1,12 +1,13 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+#include "gradient_control.h"
+#include "gradient_control_impl.h"
+
 #include "core/providers/cuda/math/binary_elementwise_ops.h"
 #include "core/providers/cuda/reduction/reduction_functions.h"
 #include "core/providers/cuda/cuda_allocator.h"
 #include "common.h"
-#include "gradient_control.h"
-#include "gradient_control_impl.h"
 
 namespace onnxruntime {
 namespace cuda {
