+++ orttraining/orttraining/training_ops/cuda/optimizer/sg.cc
@@ -1,11 +1,11 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-#include "core/providers/cuda/cuda_allocator.h"
-#include "core/providers/cuda/reduction/reduction_functions.h"
-#include "core/providers/cuda/math/binary_elementwise_ops.h"
 #include "sg.h"
 #include "sg_impl.h"
+
+#include "core/providers/cuda/reduction/reduction_functions.h"
+#include "core/providers/cuda/math/binary_elementwise_ops.h"
 
 namespace onnxruntime {
 namespace cuda {
