+++ orttraining/orttraining/training_ops/cuda/optimizer/sg_impl.cu
@@ -1,10 +1,10 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+#include "sg_impl.h"
 #include "core/providers/cuda/cuda_common.h"
 #include "core/providers/cuda/cu_inc/common.cuh"
 #include "core/providers/cuda/atomic/common.cuh"
-#include "sg_impl.h"
 
 namespace onnxruntime {
 namespace cuda {
